feat: group access request functions into same file
