+++ app/modules/aws/aws.py
@@ -3,13 +3,14 @@
 This module provides the following features:
 - Access to AWS accounts
 - Health check of AWS accounts
+- Provisioning and deprovisioning of AWS users
+- Group management (syncing, listing)
 
 """
 
 import json
 import os
 
-from server.utils import log_ops_message
 from modules.aws import aws_sso, aws_account_health, aws_access_requests
 from integrations.slack import commands as slack_commands, users as slack_users
 from integrations.aws.organizations import get_account_id_by_name
@@ -39,12 +40,30 @@ help_text = """
 
 
 def register(bot):
+    """AWS module registration.
+
+    Args:
+        bot (SlackBot): The SlackBot instance to which the module will be registered.
+    """
     bot.command(f"/{PREFIX}aws")(aws_command)
-    bot.view("aws_access_view")(access_view_handler)
+    bot.view("aws_access_view")(aws_access_requests.access_view_handler)
     bot.view("aws_health_view")(health_view_handler)
 
 
-def aws_command(ack, command, logger, respond, client, body):
+def aws_command(ack, command, logger, respond, client, body) -> None:
+    """AWS command handler.
+
+    This function handles the `/aws` command by parsing the command text and executing the appropriate action.
+
+    Args:
+        ack (function): The function to acknowledge the command.
+        command (dict): The command dictionary containing the command text.
+        logger (Logger): The logger instance.
+        respond (function): The function to respond to the command.
+        client (SlackClient): The Slack client instance.
+        body (dict): The request
+    """
+
     ack()
     logger.info("AWS command received: %s", command["text"])
 
@@ -59,7 +78,7 @@ def aws_command(ack, command, logger, respond, client, body):
         case "help" | "aide":
             respond(help_text)
         case "access":
-            request_access_modal(client, body)
+            aws_access_requests.request_access_modal(client, body)
         case "health":
             request_health_modal(client, body)
         case "user":
@@ -71,62 +90,6 @@ def aws_command(ack, command, logger, respond, client, body):
                 f"Unknown command: `{action}`. Type `/aws help` to see a list of commands.\n"
                 f"Commande inconnue: `{action}`. Tapez `/aws help` pour voir une liste des commandes."
             )
-
-
-def access_view_handler(ack, body, logger, client):
-    ack()
-
-    errors = {}
-
-    rationale = body["view"]["state"]["values"]["rationale"]["rationale"]["value"]
-
-    if len(rationale) > 2000:
-        errors["rationale"] = "Please use less than 2000 characters"
-    if len(errors) > 0:
-        ack(response_action="errors", errors=errors)
-        return
-
-    user_id = body["user"]["id"]
-
-    user = client.users_info(user=user_id)["user"]
-    email = user["profile"]["email"]
-
-    account = body["view"]["state"]["values"]["account"]["account"]["selected_option"][
-        "value"
-    ]
-
-    account_name = body["view"]["state"]["values"]["account"]["account"][
-        "selected_option"
-    ]["text"]["text"]
-
-    access_type = body["view"]["state"]["values"]["access_type"]["access_type"][
-        "selected_option"
-    ]["value"]
-
-    msg = f"<@{user_id}> ({email}) requested access to {account_name} ({account}) with {access_type} priviliges.\n\nRationale: {rationale}"
-
-    logger.info(msg)
-    log_ops_message(client, msg)
-    aws_user_id = aws_sso.get_user_id(email)
-
-    if aws_user_id is None:
-        msg = f"<@{user_id}> ({email}) is not registered with AWS SSO. Please contact your administrator.\n<@{user_id}> ({email}) n'est pas enregistré avec AWS SSO. SVP contactez votre administrateur."
-    elif expires := aws_access_requests.already_has_access(
-        account, user_id, access_type
-    ):
-        msg = f"You already have access to {account_name} ({account}) with access type {access_type}. Your access will expire in {expires} minutes."
-    elif aws_access_requests.create_aws_access_request(
-        account, account_name, user_id, email, access_type, rationale
-    ) and aws_sso.add_permissions_for_user(aws_user_id, account, access_type):
-        msg = f"Provisioning {access_type} access request for {account_name} ({account}). This can take a minute or two. Visit <https://cds-snc.awsapps.com/start#/|https://cds-snc.awsapps.com/start#/> to gain access.\nTraitement de la requête d'accès {access_type} pour le compte {account_name} ({account}) en cours. Cela peut prendre quelques minutes. Visitez <https://cds-snc.awsapps.com/start#/|https://cds-snc.awsapps.com/start#/> pour y accéder"
-    else:
-        msg = f"Failed to provision {access_type} access request for {account_name} ({account}). Please drop a note in the <#sre-and-tech-ops> channel.\nLa requête d'accès {access_type} pour {account_name} ({account}) a échouée. Envoyez une note sur le canal <#sre-and-tech-ops>"
-
-    client.chat_postEphemeral(
-        channel=user_id,
-        user=user_id,
-        text=msg,
-    )
 
 
 def health_view_handler(ack, body, logger, client):
@@ -188,87 +151,6 @@ def health_view_handler(ack, body, logger, client):
     client.views_open(
         trigger_id=body["trigger_id"],
         view=blocks,
-    )
-
-
-def request_access_modal(client, body):
-    accounts = aws_sso.get_accounts()
-    options = [
-        {
-            "text": {"type": "plain_text", "text": value},
-            "value": key,
-        }
-        for key, value in accounts.items()
-    ]
-    client.views_open(
-        trigger_id=body["trigger_id"],
-        view={
-            "type": "modal",
-            "callback_id": "aws_access_view",
-            "title": {"type": "plain_text", "text": "AWS - Account access"},
-            "submit": {"type": "plain_text", "text": "Submit"},
-            "blocks": [
-                {
-                    "block_id": "account",
-                    "type": "input",
-                    "element": {
-                        "type": "static_select",
-                        "placeholder": {
-                            "type": "plain_text",
-                            "text": "Select an account to access | Choisissez un compte à accéder",
-                        },
-                        "options": options,
-                        "action_id": "account",
-                    },
-                    "label": {"type": "plain_text", "text": "Account", "emoji": True},
-                },
-                {
-                    "block_id": "access_type",
-                    "type": "input",
-                    "label": {
-                        "type": "plain_text",
-                        "text": "What type of access do you want? :this-is-fine-fire: | Quel type d'accès désirez-vous? :this-is-fine-fire:",
-                        "emoji": True,
-                    },
-                    "element": {
-                        "type": "radio_buttons",
-                        "options": [
-                            {
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": "Read access - just need to check something \n Lecture seule - je dois juste regarder quelque chose",
-                                    "emoji": True,
-                                },
-                                "value": "read",
-                            },
-                            {
-                                "text": {
-                                    "type": "plain_text",
-                                    "text": "Write access - need to modify something \n Écriture - je dois modifier quelque chose",
-                                    "emoji": True,
-                                },
-                                "value": "write",
-                            },
-                        ],
-                        "action_id": "access_type",
-                    },
-                },
-                {
-                    "type": "input",
-                    "block_id": "rationale",
-                    "element": {
-                        "type": "plain_text_input",
-                        "multiline": True,
-                        "action_id": "rationale",
-                    },
-                    "label": {
-                        "type": "plain_text",
-                        "text": "What do you plan on doing? | Que planifiez-vous faire?",
-                        "emoji": True,
-                    },
-                },
-            ],
-        },
     )
